refactor(core): Move workflow publication outbox consumer into publication folder
